pyproject.toml
[project]
name = "modulation-explainability"
version = "0.1.0"
description = "Multi-task learning framework for automatic modulation classification with explainability"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0", 
    "numpy<2",
    "matplotlib==3.9.2",
    "wandb==0.18.3",
    "scikit-learn==1.5.2",
    "tqdm==4.66.5",
    "h5py==3.12.1",
    "pandas==2.2.3",
    "scipy==1.13.1",
    "seaborn==0.13.2",
    "pillow>=8.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []